Remove duplicate s3 code
