+++ unstructured/ingest/cli/cmds/__init__.py
@@ -22,8 +22,8 @@ from .notion import get_cmd as notion
 from .onedrive import get_cmd as onedrive
 from .outlook import get_cmd as outlook
 from .reddit import get_cmd as reddit
-from .s3_2 import get_dest_cmd as s3_dest
-from .s3_2 import get_source_cmd as s3
+from .s3 import get_dest_cmd as s3_dest
+from .s3 import get_source_cmd as s3
 from .salesforce import get_cmd as salesforce
 from .sharepoint import get_cmd as sharepoint
 from .slack import get_cmd as slack
