+++ unstructured/ingest/connector/fsspec.py
@@ -1,17 +1,20 @@
 import os
 import re
+import typing as t
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Type
 
 from unstructured.ingest.error import SourceConnectionError
-from unstructured.ingest.interfaces import (
-    BaseConnector,
+from unstructured.ingest.interfaces2 import (
     BaseConnectorConfig,
+    BaseDestinationConnector,
     BaseIngestDoc,
-    ConnectorCleanupMixin,
+    BaseSourceConnector,
     IngestDocCleanupMixin,
-    StandardConnectorConfig,
+    PartitionConfig,
+    ReadConfig,
+    SourceConnectorCleanupMixin,
+    WriteConfig,
 )
 from unstructured.ingest.logger import logger
 
@@ -31,7 +34,7 @@ SUPPORTED_REMOTE_FSSPEC_PROTOCOLS = [
 class SimpleFsspecConfig(BaseConnectorConfig):
     # fsspec specific options
     path: str
-    recursive: bool
+    recursive: bool = False
     access_kwargs: dict = field(default_factory=dict)
     protocol: str = field(init=False)
     path_without_protocol: str = field(init=False)
@@ -69,9 +72,6 @@ class SimpleFsspecConfig(BaseConnectorConfig):
         self.dir_path = match.group(1)
         self.file_path = match.group(2) or ""
 
-    def get_access_kwargs(self) -> dict:
-        return self.access_kwargs
-
 
 @dataclass
 class FsspecIngestDoc(IngestDocCleanupMixin, BaseIngestDoc):
@@ -82,20 +82,21 @@ class FsspecIngestDoc(IngestDocCleanupMixin, BaseIngestDoc):
     method is not called, the file is left behind on the filesystem to assist debugging.
     """
 
-    config: SimpleFsspecConfig
+    connector_config: SimpleFsspecConfig
     remote_file_path: str
 
     def _tmp_download_file(self):
-        return Path(self.standard_config.download_dir) / self.remote_file_path.replace(
-            f"{self.config.dir_path}/",
+        download_dir = self.read_config.download_dir if self.read_config.download_dir else ""
+        return Path(download_dir) / self.remote_file_path.replace(
+            f"{self.connector_config.dir_path}/",
             "",
         )
 
     @property
     def _output_filename(self):
         return (
-            Path(self.standard_config.output_dir)
-            / f"{self.remote_file_path.replace(f'{self.config.dir_path}/', '')}.json"
+            Path(self.partition_config.output_dir)
+            / f"{self.remote_file_path.replace(f'{self.connector_config.dir_path}/', '')}.json"
         )
 
     def _create_full_tmp_dir_path(self):
@@ -109,8 +110,8 @@ class FsspecIngestDoc(IngestDocCleanupMixin, BaseIngestDoc):
         from fsspec import AbstractFileSystem, get_filesystem_class
 
         self._create_full_tmp_dir_path()
-        fs: AbstractFileSystem = get_filesystem_class(self.config.protocol)(
-            **self.config.get_access_kwargs(),
+        fs: AbstractFileSystem = get_filesystem_class(self.connector_config.protocol)(
+            **self.connector_config.access_kwargs,
         )
         logger.debug(f"Fetching {self} - PID: {os.getpid()}")
         fs.get(rpath=self.remote_file_path, lpath=self._tmp_download_file().as_posix())
@@ -121,40 +122,45 @@ class FsspecIngestDoc(IngestDocCleanupMixin, BaseIngestDoc):
         return self._tmp_download_file()
 
 
-class FsspecConnector(ConnectorCleanupMixin, BaseConnector):
+class FsspecSourceConnector(SourceConnectorCleanupMixin, BaseSourceConnector):
     """Objects of this class support fetching document(s) from"""
 
-    config: SimpleFsspecConfig
-    ingest_doc_cls: Type[FsspecIngestDoc] = FsspecIngestDoc
+    connector_config: SimpleFsspecConfig
+    ingest_doc_cls: t.Type[FsspecIngestDoc] = FsspecIngestDoc
 
     def __init__(
         self,
-        standard_config: StandardConnectorConfig,
-        config: SimpleFsspecConfig,
+        read_config: ReadConfig,
+        connector_config: BaseConnectorConfig,
+        partition_config: PartitionConfig,
     ):
         from fsspec import AbstractFileSystem, get_filesystem_class
 
-        super().__init__(standard_config, config)
-        self.fs: AbstractFileSystem = get_filesystem_class(self.config.protocol)(
-            **self.config.get_access_kwargs(),
+        super().__init__(
+            read_config=read_config,
+            connector_config=connector_config,
+            partition_config=partition_config,
+        )
+        self.fs: AbstractFileSystem = get_filesystem_class(self.connector_config.protocol)(
+            **self.connector_config.access_kwargs,
         )
 
     def initialize(self):
         """Verify that can get metadata for an object, validates connections info."""
-        ls_output = self.fs.ls(self.config.path_without_protocol)
+        ls_output = self.fs.ls(self.connector_config.path_without_protocol)
         if len(ls_output) < 1:
             raise ValueError(
-                f"No objects found in {self.config.path}.",
+                f"No objects found in {self.connector_config.path}.",
             )
 
     def _list_files(self):
-        if not self.config.recursive:
+        if not self.connector_config.recursive:
             # fs.ls does not walk directories
             # directories that are listed in cloud storage can cause problems
             # because they are seen as 0 byte files
             return [
                 x.get("name")
-                for x in self.fs.ls(self.config.path_without_protocol, detail=True)
+                for x in self.fs.ls(self.connector_config.path_without_protocol, detail=True)
                 if x.get("size") > 0
             ]
         else:
@@ -163,7 +169,7 @@ class FsspecConnector(ConnectorCleanupMixin, BaseConnector):
             return [
                 k
                 for k, v in self.fs.find(
-                    self.config.path_without_protocol,
+                    self.connector_config.path_without_protocol,
                     detail=True,
                 ).items()
                 if v.get("size") > 0
@@ -172,9 +178,49 @@ class FsspecConnector(ConnectorCleanupMixin, BaseConnector):
     def get_ingest_docs(self):
         return [
             self.ingest_doc_cls(
-                standard_config=self.standard_config,
-                config=self.config,
+                read_config=self.read_config,
+                connector_config=self.connector_config,
+                partition_config=self.partition_config,
                 remote_file_path=file,
             )
             for file in self._list_files()
         ]
+
+
+class FsspecDestinationConnector(BaseDestinationConnector):
+    connector_config: SimpleFsspecConfig
+
+    def __init__(self, write_config: WriteConfig, connector_config: BaseConnectorConfig):
+        from fsspec import AbstractFileSystem, get_filesystem_class
+
+        super().__init__(write_config=write_config, connector_config=connector_config)
+        self.fs: AbstractFileSystem = get_filesystem_class(self.connector_config.protocol)(
+            **self.connector_config.access_kwargs,
+        )
+
+    def initialize(self):
+        pass
+
+    def write(self, docs: t.List[BaseIngestDoc]) -> None:
+        from fsspec import AbstractFileSystem, get_filesystem_class
+
+        fs: AbstractFileSystem = get_filesystem_class(self.connector_config.protocol)(
+            **self.connector_config.access_kwargs,
+        )
+
+        logger.info(f"Writing content using filesystem: {type(fs).__name__}")
+
+        for doc in docs:
+            s3_file_path = str(doc._output_filename).replace(
+                doc.partition_config.output_dir,
+                self.connector_config.path,
+            )
+            s3_folder = self.connector_config.path
+            if s3_folder[-1] != "/":
+                s3_folder = f"{s3_file_path}/"
+            if s3_file_path[0] == "/":
+                s3_file_path = s3_file_path[1:]
+
+            s3_output_path = s3_folder + s3_file_path
+            logger.debug(f"Uploading {doc._output_filename} -> {s3_output_path}")
+            fs.put_file(lpath=doc._output_filename, rpath=s3_output_path)
