+++ unstructured/ingest/connector/s3.py
@@ -2,12 +2,17 @@ from dataclasses import dataclass
 from typing import Type
 
 from unstructured.ingest.connector.fsspec import (
-    FsspecConnector,
+    FsspecDestinationConnector,
     FsspecIngestDoc,
+    FsspecSourceConnector,
     SimpleFsspecConfig,
 )
-from unstructured.ingest.error import SourceConnectionError
-from unstructured.ingest.interfaces import StandardConnectorConfig
+from unstructured.ingest.interfaces2 import (
+    BaseConnectorConfig,
+    PartitionConfig,
+    ReadConfig,
+    WriteConfig,
+)
 from unstructured.utils import requires_dependencies
 
 
@@ -21,19 +26,29 @@ class S3IngestDoc(FsspecIngestDoc):
     remote_file_path: str
     registry_name: str = "s3"
 
-    @SourceConnectionError.wrap
     @requires_dependencies(["s3fs", "fsspec"], extras="s3")
     def get_file(self):
         super().get_file()
 
 
-@requires_dependencies(["s3fs", "fsspec"], extras="s3")
-class S3Connector(FsspecConnector):
+class S3SourceConnector(FsspecSourceConnector):
     ingest_doc_cls: Type[S3IngestDoc] = S3IngestDoc
 
+    @requires_dependencies(["s3fs", "fsspec"], extras="s3")
     def __init__(
         self,
-        config: SimpleS3Config,
-        standard_config: StandardConnectorConfig,
-    ) -> None:
-        super().__init__(standard_config, config)
+        read_config: ReadConfig,
+        connector_config: BaseConnectorConfig,
+        partition_config: PartitionConfig,
+    ):
+        super().__init__(
+            read_config=read_config,
+            connector_config=connector_config,
+            partition_config=partition_config,
+        )
+
+
+class S3DestinationConnector(FsspecDestinationConnector):
+    @requires_dependencies(["s3fs", "fsspec"], extras="s3")
+    def __init__(self, write_config: WriteConfig, connector_config: BaseConnectorConfig):
+        super().__init__(write_config=write_config, connector_config=connector_config)
