+++ unstructured/ingest/runner/__init__.py
@@ -50,5 +50,4 @@ __all__ = [
     "sharepoint",
     "slack",
     "wikipedia",
-    "writer_map",
 ]
